refactor(assumptions): migrate order handlers to new design
Migrate Q.zero
