+++ sympy/assumptions/ask_generated.py
@@ -107,6 +107,7 @@ def get_known_facts_cnf():
         Q.finite | ~Q.transcendental,
         Q.hermitian | ~Q.real,
         Q.nonnegative | ~Q.positive,
+        Q.nonnegative | ~Q.zero,
         Q.nonzero | ~Q.negative,
         Q.nonzero | ~Q.positive,
         Q.rational | ~Q.integer,
@@ -114,31 +115,31 @@ def get_known_facts_cnf():
         Q.real | ~Q.negative,
         Q.real | ~Q.positive,
         Q.real | ~Q.rational,
+        Q.real | ~Q.zero,
         Q.integer | ~Q.even,
         Q.integer | ~Q.odd,
         Q.integer | ~Q.prime,
-        Q.nonnegative | ~Q.zero,
         ~Q.algebraic | ~Q.transcendental,
         ~Q.antihermitian | ~Q.hermitian,
         ~Q.finite | ~Q.infinite,
         ~Q.imaginary | ~Q.real,
         ~Q.irrational | ~Q.rational,
         ~Q.negative | ~Q.positive,
+        ~Q.negative | ~Q.zero,
         ~Q.negative | Q.nonpositive,
+        ~Q.positive | ~Q.zero,
+        ~Q.zero | Q.even,
+        ~Q.zero | Q.nonpositive,
         Q.positive | ~Q.prime,
-        Q.real | ~Q.zero,
         Q.invertible | Q.singular,
         Q.infinite | Q.real | ~Q.extended_real,
         Q.negative | Q.positive | ~Q.nonzero,
-        ~Q.negative | ~Q.zero,
-        ~Q.positive | ~Q.zero,
+        Q.positive | Q.zero | ~Q.nonnegative,
         Q.complex_elements | ~Q.real_elements,
-        Q.even | ~Q.zero,
         Q.fullrank | ~Q.invertible,
         Q.invertible | ~Q.positive_definite,
         Q.invertible | ~Q.unitary,
         Q.lower_triangular | ~Q.diagonal,
-        Q.nonpositive | ~Q.zero,
         Q.normal | ~Q.diagonal,
         Q.normal | ~Q.unitary,
         Q.positive_definite | ~Q.orthogonal,
@@ -152,15 +153,14 @@ def get_known_facts_cnf():
         Q.triangular | ~Q.upper_triangular,
         Q.unitary | ~Q.orthogonal,
         Q.upper_triangular | ~Q.diagonal,
-        Q.positive | ~Q.nonnegative | Q.zero,
+        Q.negative | Q.zero | ~Q.nonpositive,
+        Q.negative | Q.positive | Q.zero | ~Q.real,
         ~Q.composite | ~Q.prime,
         ~Q.even | ~Q.odd,
         ~Q.invertible | ~Q.singular,
-        Q.negative | Q.zero | ~Q.nonpositive,
         ~Q.integer | Q.even | Q.odd,
         Q.algebraic | Q.transcendental | ~Q.complex | ~Q.finite,
         Q.irrational | Q.rational | ~Q.finite | ~Q.real,
-        Q.negative | Q.positive | ~Q.real | Q.zero,
         ~Q.real | Q.orthogonal | ~Q.unitary,
         Q.lower_triangular | Q.upper_triangular | ~Q.triangular,
         Q.diagonal | ~Q.lower_triangular | ~Q.upper_triangular,
